+++ netbox_kea/tests/utils.py
@@ -33,7 +33,7 @@ def _make_db_server(**kwargs) -> Server:
     """
     defaults = {
         "name": "test-kea",
-        "server_url": "https://kea.example.com",
+        "ca_url": "https://kea.example.com",
         "dhcp4": True,
         "dhcp6": True,
         "has_control_agent": True,